% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_runpar.R
\name{plot_runpar}
\alias{plot_runpar}
\title{Plot the parameters for a simulation run}
\usage{
plot_runpar(df, main = "Run Parameters", featurelist, exclude_items = NA)
}
\arguments{
\item{df}{Either a list of parameters, or a dataframe containing
the run parameters to plot on the screen.  If a list,
each item should be named as the names correspond to the
parameter name.  If a dataframe, the row names should
be the parameter names.}

\item{main}{The main title at the top panel of the screen.  If no
title is desired, then set to an empty string "".}

\item{featurelist}{A list of feature vectors to plot on the bottom panel.}

\item{exclude_items}{A vector of row numbers or row names corresponding
to the rows to exclude from showing in the top panel.}
}
\description{
Plots the parameters for a simulation in its own plotting window.
This is normally used as a header page to describe the key
parameters for a simulation.  The function provides
}

