% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_features.R
\name{RMSE}
\alias{RMSE}
\title{Calculate the Root Mean Square Error of a 2 column matrix}
\usage{
RMSE(mat)
}
\arguments{
\item{mat}{The two column matrix used to calculate the mean squared difference.}
}
\value{
Returns the RMS error of the two columns in the matrix.
}
\description{
Given a two column matrix (typically yhat and y), this function
calculates the root mean square error between y and yhat.
}
\details{
ssuming N is the number or rows, the calculation for the RMSE
value is done according to the following formula:

\deqn{\sqrt \sum(yhat - y)^2 / N}
}

